initial attempt at refactoring the build config
