+++ buildSrc/src/main/kotlin/ijrust/conventions/base.gradle.kts
@@ -1,0 +1,16 @@
+package ijrust.conventions
+
+import ijrust.IJRustBuildProperties
+
+plugins {
+    base
+//    idea
+}
+
+extensions.create<IJRustBuildProperties>("ijRustBuild")
+//
+//idea {
+//    module {
+//        generatedSourceDirs.add(file("src/gen"))
+//    }
+//}
