+++ duplicates/build.gradle.kts
@@ -1,0 +1,8 @@
+plugins {
+    ijrust.conventions.intellij
+}
+
+dependencies {
+    implementation(project(":"))
+    testImplementation(project(":", "testOutput"))
+}
