+++ gradle.properties
@@ -22,6 +22,8 @@ org.gradle.jvmargs=-Xmx3g
 # Enable Gradle Build Cache. It is also configured in settings.gradle.kts.
 org.gradle.caching=true
 
+org.gradle.logging.stacktrace=all
+
 systemProp.org.gradle.internal.repository.max.tentatives=10
 systemProp.org.gradle.internal.http.connectionTimeout=120000
 systemProp.org.gradle.internal.http.socketTimeout=120000
@@ -34,3 +36,19 @@ kotlin.stdlib.default.dependency=false
 # See https://plugins.jetbrains.com/docs/intellij/using-kotlin.html#incremental-compilation
 # and https://youtrack.jetbrains.com/issue/KT-57757 for more details
 kotlin.incremental.useClasspathSnapshot=false
+
+
+# Existent IDE versions can be found in the following repos:
+# https://www.jetbrains.com/intellij-repository/releases/
+# https://www.jetbrains.com/intellij-repository/snapshots/
+ideaVersion=IU-2023.1
+clionVersion=CL-2023.1
+
+# https://plugins.jetbrains.com/plugin/12775-native-debugging-support/versions
+nativeDebugPlugin=com.intellij.nativeDebug:231.8109.91
+# https://plugins.jetbrains.com/plugin/227-psiviewer/versions
+psiViewerPlugin=PsiViewer:231-SNAPSHOT
+
+# please see https://www.jetbrains.org/intellij/sdk/docs/basics/getting_started/build_number_ranges.html for description
+sinceBuild=231.7515
+untilBuild=231.*
