+++ grammar-kit-fake-psi-deps/build.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+    ijrust.conventions.intellij
+}
