+++ grazie/build.gradle.kts
@@ -1,0 +1,12 @@
+plugins {
+    ijrust.conventions.intellij
+}
+
+intellij {
+    plugins.set(listOf(ijRustBuild.graziePlugin))
+}
+
+dependencies {
+    implementation(project(":"))
+    testImplementation(project(":", "testOutput"))
+}
