+++ intelliLang/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+    ijrust.conventions.intellij
+}
+
+intellij {
+    plugins.set(listOf(ijRustBuild.intelliLangPlugin))
+}
+dependencies {
+    implementation(project(":"))
+    testImplementation(project(":", "testOutput"))
+}
