+++ js/build.gradle.kts
@@ -1,0 +1,12 @@
+plugins {
+    ijrust.conventions.intellij
+}
+
+intellij {
+    plugins.set(listOf(ijRustBuild.javaScriptPlugin))
+}
+
+dependencies {
+    implementation(project(":"))
+    testImplementation(project(":", "testOutput"))
+}
