+++ profiler/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    ijrust.conventions.intellij
+}
+
+intellij {
+    version.set(ijRustBuild.clionVersion)
+    plugins.set(ijRustBuild.clionPlugins)
+}
+
+dependencies {
+    implementation(project(":"))
+    testImplementation(project(":", "testOutput"))
+}
