+++ settings.gradle.kts
@@ -1,19 +1,42 @@
 rootProject.name = "intellij-rust"
-// Special module with run, build and publish tasks
-include("plugin")
+
+pluginManagement {
+    repositories {
+        mavenCentral()
+        maven("https://oss.sonatype.org/content/repositories/snapshots/") {
+            mavenContent { snapshotsOnly() }
+        }
+        gradlePluginPortal()
+    }
+}
+
+@Suppress("UnstableApiUsage")
+dependencyResolutionManagement {
+//    repositoriesMode.set(RepositoriesMode.PREFER_SETTINGS)
+
+    repositories {
+        mavenCentral()
+        maven("https://cache-redirector.jetbrains.com/repo.maven.apache.org/maven2")
+        maven("https://cache-redirector.jetbrains.com/intellij-dependencies")
+    }
+}
+
+
 include(
-    "grammar-kit-fake-psi-deps",
-    "idea",
-    "clion",
-    "debugger",
-    "profiler",
-    "copyright",
-    "coverage",
-    "intelliLang",
-    "duplicates",
-    "grazie",
-    "js",
-    "ml-completion"
+    ":plugin", // Special module with run, build and publish tasks
+
+    ":grammar-kit-fake-psi-deps",
+    ":idea",
+    ":clion",
+    ":debugger",
+    ":profiler",
+    ":copyright",
+    ":coverage",
+    ":intelliLang",
+    ":duplicates",
+    ":grazie",
+    ":js",
+    ":ml-completion"
 )
 
 // Configure Gradle Build Cache. It is enabled in `gradle.properties` via `org.gradle.caching`.
@@ -25,9 +48,5 @@ buildCache {
     }
 }
 
-pluginManagement {
-    repositories {
-        maven("https://oss.sonatype.org/content/repositories/snapshots/")
-        gradlePluginPortal()
-    }
-}
+enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
+enableFeaturePreview("STABLE_CONFIGURATION_CACHE")
